+++ RevitAddin.WpfUi.Sample/Revit/App.cs
@@ -21,17 +21,6 @@ namespace RevitAddin.WpfUi.Sample.Revit
             ribbonPanel.CreatePushButton<Commands.CommandSimple>("Simple")
                 .SetLargeImage("Resources/Revit.ico");
 
-            var themesDictionary = new Wpf.Ui.Markup.ThemesDictionary();
-            var controlsDictionary = new Wpf.Ui.Markup.ControlsDictionary();
-
-            ResourcesDictionaries.Add(themesDictionary);
-            ResourcesDictionaries.Add(controlsDictionary);
-
-            foreach (var item in ResourcesDictionaries)
-            {
-                //Application.Current?.Resources.MergedDictionaries.Add(item);
-            }
-
             return Result.Succeeded;
         }
 
@@ -40,11 +29,6 @@ namespace RevitAddin.WpfUi.Sample.Revit
         public Result OnShutdown(UIControlledApplication application)
         {
             ribbonPanel?.Remove();
-
-            foreach (var item in ResourcesDictionaries)
-            {
-                Application.Current?.Resources.MergedDictionaries.Remove(item);
-            }
 
             return Result.Succeeded;
         }